\name{airway}
\alias{airway}
\docType{data}
\title{Airway smooth muscle cells}
\description{Read counts per gene for airway smooth muscle cell lines RNA-Seq experiment}
\details{
  For details on the gene model and read counting procedure, see the
  package vignette. 

  The dataset contains four cell lines in two conditions: control and
  treatment with dexamethasone.
}
\usage{
data("airway")
}
\format{RangedSummarizedExperiment}
\source{FASTQ files from SRA, phenotypic data from GEO}
\references{Himes BE, Jiang X, Wagner P, Hu R, Wang Q, Klanderman B,
  Whitaker RM, Duan Q, Lasky-Su J, Nikolos C, Jester W, Johnson M,
  Panettieri R Jr, Tantisira KG, Weiss ST, Lu Q. 'RNA-Seq Transcriptome
  Profiling Identifies CRISPLD2 as a Glucocorticoid Responsive Gene that
  Modulates Cytokine Function in Airway Smooth Muscle Cells.' PLoS
  One. 2014 Jun 13;9(6):e99625. PMID: 24926665. GEO: GSE52778.
}
\keyword{datasets}
